fix(product-attributes): move default color term seeding to admin-only path
- Remove woocommerce_attribute_added hook and wrapper from VisualAttributeTermAdmin.
- Move seed_visual_attribute_terms() and default colors from VisualAttributeTermMeta to VisualAttributeTermAdmin.
- Call seeder from WC_Admin_Attributes::process_add_attribute() after wc_create_attribute succeeds.
- Validate taxonomy against the created attribute ID before inserting terms.
- Move seeding tests to VisualAttributeTermAdminTest and use deterministic unique suffixes.

Addresses PR review feedback.

Refs #65923
